Completely refactor how methods are tracked.
- We now reuse tags, when possible. \o/
- We now re-emit methods in the class, when their arguments or return type change (except we don't yet for virtual methods).
- We have deterministic ordering of methods. \o/

Internally, the ImplementedClassMethod class keeps track of each emitted method,
with the methods it hides.

This is used as the basis for dispatching to the correct emitters for the actual
implementation.

The actual implementation now decides if an accessor method is to be emitted,
or if the old accessor method is sufficient.